Working Prototype Version
Added a significant amount of initial content for the prototype. A lot of the code is subject to change considering I know ways to clean it up more. This is a rough draft of the game, so things are still quite messy. One major bug is that the player can walk off screen. That can be fixed later.